requirements.txt
python-telegram-bot==13.7
woocommerce
qrcode
pandas
python-dotenv
babel
datetime
openpyxl